chore: changes required in order to deploy with internal tools (#36)
Align a few dependencies to the library versions available in the Amazon internal repositories so we can deploy construct hub.

Additionally, move tests under `src/__tests__` so that they are executed as JavaScript and not TypeScript. This is required in order for `NodejsFunction` to be able to find the entry point.