chore: pre-bundle lambda functions
The internal Amazon build system makes it extremely hard to use `esbuild` to bundle during app synthesis. So we are taking this opportunity to experiment with the idea of pre-bundling lambda functions during library build, alongside synthesis of a construct class that can be used seamlessly in the construct hub code.

The developer experience is pretty neat:

1. Create a file somewhere under the `src` tree with the `.lambda.ts` extension.
2. This file should export a single async function called `handler`.
3. Run `projen`.
4. This will do the following:
   1. Add a compile task that uses `esbuild` to create a bundle from them under `lib/xxx.bundle/index.js`
   2. Generate TypeScript code under `xxx.ts` with a construct that extends `lambda.Function` which points to the bundled asset.

Now, in your CDK code, just import the generated construct and use it.