chore: turn construct-hub-webapp to a dev dependency
Since construct-hub-webapp currently takes a normal dependency on react-related libraries, bundling it using `bundledDeps` means we will bundle dozens of modules and the tarball size gets to 30MiB.

Instead, take a dev-dependency on it and just copy the `build/` directory from construct-hub-webapp to `website/`, which is then used as the asset.

We can add support in projen to change react-related deps to be dev-deps but this is a good solution for now.