feat(monitoring): web canaries & alarm actions
Add support for defining web canaries (http pings) and setting actions for high-severity and normal-severity alarms.

To that end, users are now required to pass in at least the ARN of a CloudWatch action to be triggered in response to high-severity alarms. Optionally, an action can also be specified for normal severity alarms.

Additionally, we no long auto-monitor the entire stack, but require that various elements in the system add themselves to be watched through the `Watchful` API (e.g. `monitoring.watchful.watchDynamoDbTable()`).