feat(monitoring): web canaries & alarm actions (#53)
Add support for defining web canaries (HTTP pings) and setting actions for high-severity and normal-severity alarms.

To that end, users are now required to pass in at least the ARN of a CloudWatch action to be triggered in response to high-severity alarms. Optionally, an action can also be specified for normal severity alarms.

Note that many of the potentially configurable options are currently not exposed. For example, canaries always run every 1min, sum errors within a period of 5min and alarm if there are 4 consecutive failures. We will expose additional configuration as needed.

Additionally, we no long auto-monitor the entire stack, but require that various elements in the system add themselves to be watched through the `Watchful` API (e.g. `monitoring.watchful.watchDynamoDbTable()`).

Project changes: allow lambda code to take a dev-dependency since it is bundled; add a `bundle` task which bundles all lambda functions.

*Why didn’t we use CloudWatch Synthetics?* because I couldn’t find a way to pass values into the test handler. If anyone knows of a way, I’d be delighted to switch this up later.