feat: transliteration function
When a `.tgz` package is uploaded into an S3 bucket, the `jsii-rosetta
transpile` function is automatically run on the assembly contained
therein.

When sucessful, the transliterated assemblies are automatically uploaded
to the same S3 bucket, with a different object key.

The initial implementation only performs Python generation (this is
hard-coded and will be changed later).