fix: S3 update notifications have URI-encoded keys
The key in the S3 update notification is URI-encoded such that the `@`
in scoped packages' keys is represented as `%40`. This currently fails
the key check and causes scoped packages to be impossible to process.

This addresses the problem by allowing `%40` in lieu of `@`, and
handling the scope somewhat more specifically in the function.