feat: Catalog builder function
This function can be invoked periodically or through an event to
re-build the `catalog.json` object in the package store S3 bucket. It
can either work incrementally (making changes to the existing object),
or from scratch (re-building the catalog object from the complete list
of indexed packages as found in the S3 bucket, which may be slower).

Fixes #6