fix(discovery): invalid header when passing error message (#99)
When an error occurs during processing we attempt to save the faulty input to s3. For debugging purpose we save the error in the S3 object metadata. Since the `Metadata` keys are passed as http header, if the error contains invalid header characters (such as `\n` which are prevalent in stack trace) it will fail to put the object with:

```
TypeError [ERR_INVALID_CHAR]: Invalid character in header content ["x-amz-meta-Error"]
```

We can sanitize the error, but since we kill the function execution if this hail-mary write fails we should play it safe, and not use the headers. Instead, add a special  key with the error as the value to the faulty object. This also means that we are not limited by the size of the allowed S3 user defined metadata, and can save the full error.