chore(discovery): smaller batch size (#112)
Changing the batch size used when issuing a request to `couchdb`. This is done to reduce the memory size required during the function execution. Since smaller batch size means lower (average) batch processing time, this PR also reduce the time we require to remain in the Lambda execution in order to request an additional batch.

As mentioned in the code, checking for the time left in the function execution is a best effort, as the distribution of the time it takes to process a package/batch is non uniform, thus we expect the function to time out in some invocations and rely on the downstream idempotency.