feat(discovery): decouple npmjs follower & staging
Separating the npmjs follower from staging allows for more efficient
tracking against the npmjs catalog, as the downloading and staging of
tarballs is parallelized more efficiently in multiple processes instead
of a single node process' libuv thread pool.

The two functions are linked together using an SQS queue, which serves
as a batching mechanism, and the staging function has a
dead-letter-queue so that "poison pill" messages are effectively
side-lined for manual review.

The new npmjs follower maintains state about which versions it has
already processed in the marker file (which auto-converts from the old
to the new format), so that it avoids re-notifying about package
versions that have not changed since they were last staged. It notifies
about ALL new package versions, not just the latest one, as was
previously the case, so that the Construct Hub will injest ALL
compatible and relevant versions of packages as part of a back-fill.