fix: catalog builder triggered before metadata is available (#129)
The catalog builder is invoked when an assembly file is uploaded to the bucket. It then proceeds to download the `metadata.json` file for that assembly. Problem is that the `metadata.json` and `assembly.json` files are uploaded in parallel.

This PR makes it so the ingestion function uploads the assembly file **last**, only after the `metadata.json`.

Still in draft since it requires some more testing. But the approach holds true.