feat: filter packages based on license
Only select packages that meet the relevant criteria AND also feature an
eligible license (`Apache`, `BSD`, or `MIT`). The check is performed
both during discovery (based on what's in `package.json`) and during
ingestion (based on what's in the `.jsii` assembly). The discovery
function reports a metric with how many package-versions it has ignored
due to their not passing the license check.

We currently don't match the license ID with the contents of a `LICENSE`
file (if one is present), but the content of that license file is copied
into the `metadata.json` file so the front-end may render it if needed.

Fixes #145