feat: ignore packages with name or version mismatch
There is one package (`@rhaws/aws-ec2`) that is a re-packaging of
another (`@aws-cdk/aws-ec2`) with a different package name, however the
`.jsii` assembly included in said package still carries the original
package name, which causes undesired confusion down the processing
pipeline.

This rejects packages if the name and version found in `package.json` do
not correspond to those found in the `.jsii` assembly.