chore: split transliterator into a function per language (#178)
Previously, we had a single function that generated docs for all supported languages. 
When a failure in a specific language happens, the function would log that failure and proceed to the next language. 

While we do want that behavior, we also want to have good visibility into those failures. Splitting the transliterator to a separate lambda per language gives us that by default because now the invocations represent true failures/successes. 

In addition, this splitting also performs better since doc generation happens in parallel for each language.  

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*