feat(transliterator): write marker objects for un-supported languages
When a given assembbly does not support a particular language, write
a marker object to note the process has run, but the language is not
supported.

This allows for accurate counting of which objects are missing in S3 due
to not having been generated at all, instead of also counting those that
simply cannot exist due to not being supported.

Fixes #186

# Conflicts:
#	src/__tests__/__snapshots__/construct-hub.test.ts.snap