chore: add Cache-Control headers to S3 objects
The objects served by CloudFront can use Cache-Control headers
to finely control the caching behavior. For most objects, we
simply specify they are public, making it clear CloudFront is
allowed to retain them in cache (this makes the implicit default
explicit). For the catalog.json object, also specify a max-age of
5 minutes, so that CloudFront will somewhat frequently refresh
it's cache entries.