chore: improved StepFunctions orchestration (#224)
This new workflow enables better tracing of what happened, and where,
especially when it comes to troubleshooting messages from the dead
letter queue.

It adds new packages to `catalog.json` only once, if either the TypeScript
or Python documentation generation succeeded (and not otherwise), and
sends the input to the DLQ if either of them failed (not necessarily both).

It tags the input with the SFN execution information, so it is easy to get back
to that from a message in the DLQ.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*