fix: `ConstructHubProps.isolateLambdas` default incorrect
A bug was causing the default for `isolateLambdas` to be `false` instead
of `true`, resulting in deployments having NAT gateways by default,
which is not the safest option.

Added a snapshot test covering the explicit disabling of
`isolateLambdas`, so that inadvertent changes to this default behavior
are caught in the future.