fix: better error handling in step functions
The `Catch` blocks were previously assuming all error causes are always
exposed as an "escaped JSON string", however this is only true for
certain causes (i.e: Lambda invoke errors).

This changes the catch logic so that it does not attempt to JSON-parse
the data that is not guaranteed to be an escaped JSON string, so the
state machines don't fail in this case.