chore: tune sensitivity of discovery alarm
Instead of alarming on 2 breaching point with a 15 minutes interval (so
alarm after 30 minutes), alarm on 3 breaching points with 5 minutes
interval (alarming after 15 minutes, with better specificity).

This reduces the risk of false positives induced by unavailable data.
Also changed the `Errors` alarm to treat missing data as... missing (so
they are not accounted for in evaluating the metric), and the
`Invocations` alarm treats missing data as breaching (because no metric
is emitted if the function does not get invoked).