feat: allow configuring multiple package sources
Replaced the `Discovery` function with a pluggable `IPackageSource` API,
and providing an `NpmJs` as well as `CodeArtifact` source. These sources
can optionally be passed into the `ConstructHub` constructor to register
alternate package sources (by default, an `NpmJs` source is registered).

This unfortunately involved relativelu wide-spread changes in the
codebase, as some features had to be made visible from the public API of
the module, and some code needed to be re-used (hence, refactored to a
different location).

Fixes #281