feat: allow specific custom links
Allows contruct-hub admins to specify a list of allowed links that
package authors can include in their package.json to be shown on their
package's details page. For Example:

A package author can add the following to their package.json
```json
"constructHub": {
  "links": {
    "SLA": "https://link.to.my.support.info",
    "Author": "mailto:me@you.com"
  }
}
```

Then, the administrator of construct hub can add these to an allow list
like so:
```ts
new ConstructHub(this, "ConstructHub", {
  ...myProps,
  packageLinks: [{
    name: 'Service Level Agreement',
    value: 'SLA',
  }, {
    name: 'Contact',
    value: 'Author',
    displayText: 'Email Me!',
  }]
});
```

With these links allowed, the ingestion lambda will search for these
values and place them in a package's metadata.json file. Additionally,
the configuration is put into a `config.json` file alongside the webapp
assets to be loaded by the frontend in order to determine what links to
render.