chore: tweak retry behavior in Orchestration
Significantly increase the total amount of time covered by retry
attempts around Lambda function throttles (allowing up to three hours in
total).

This is in interim until we can vet a better back-pressure management
solution.

This also stops attempting to retry doc-gen for packages that fail due
to running out of disk space, as those errors are unlikely to be
transient.