feat: computed package tags
Adds support for adding custom tags to packages when relevant. Users can
specify the label and color of a tag, and to which packages it is
applied.

```ts
new ConstructHub(this, "ConstructHub", {
  ...myProps,
  packageTags: [{
    label: 'Official',
    color: '#00FF00',
    condition: TagCondition.eqls(['name'], 'construct-hub'),
  }]
});
```

The above results in packages with the `name` of `construct-hub` receiving the `Official` tag, which is colored green. Tags can be used in search or just as labels.

`TagCondition` supports the following logical operators, but more can be
added as needed.
```
TagCondition.and(...conditions) // &&
TagCondition.or(...conditions) // ||
TagCondition.not(condition) // !
TagCondition.eqls(key, value) // ===
```