chore: set embedded metrics environment to "Local"
This disables all automated decoration (i.e: dimensions), which we are
not using anyway, and ensures metrics are flushed out to STDOUT. This is
done by setting an environment variable on Lambda functions that
controls the default behavior of the `aws-embedded-metrics` library.

Locally, this has the effect of correctly emitting the metrics within
the canary lambda function, which stopped working in production a couple
of days/weeks ago for an unknown reason.