fix: NpmJs not registering new packages
The CouchDB replica for the npmjs.com registry has been rotated,
resulting in resetting the sequence numbers. The NpmJS follower did not
account for this possibility and is hence not registering any changes,
as the marker is now in the distant future from the new sequence.

Added an alarm (low severity) in case no NpmJS changes are registered
for 2 intervals (10 minutes), so we are warned about this in the future.

Additionally, made the NpmJS Follower detect sequence number resets, and
start crawling back from `0` again in this case.

Ref: https://github.blog/changelog/2021-09-08-npm-couchdb-upgrade-will-reset-sequence-number/