fix(ingestion): invalid version when encountering SemVer ranges
The construct framework detection logic caused packages to fail
ingesting when they declared a range dependency on the construct
framework package (as opposed to a version-pinned dependency).
This changes the logic to parse the ranges and select the minimum
supported major version instead, which corrects the issue.

Additionally, the framework detection only looked into dependencies,
meaning that the framework packages themselves (i.e: `@aws-cdk/core`)
did not get a framework identified. This is now fixed.

Fixes #376