feat(orchestration): SFN-based re-drive workflow
Re-implement the "re-generate all documentation" as a StepFunctions
state machine, so it does not need to fit within a 15 minutes lambda
function.

It had to be implemented as two different state machines in order to
avoid hitting the 25K events per execution limit, and could be further
modified to avoid proliferating events in the S3 list page traversal (by
fetching the next page in a separate execution, instead of looping), but
this does not appear to be necessary at this stage (we can grow to 5
times more pages before we start getting close to the 25K limit again).