feat: package stats fetching (#385)
Adds a "package stats" component to the backend that runs a Lambda once a day that generates/updates the `stats.json` file in the package data S3 bucket. The file is formatted as shown:

```
{
  "packages": {
    "@foo/bar": {
      "downloads": {
        "npm": 10
      }
    },
    "unscoped-pkg": {
      "downloads": {
        "npm": 123
      }
    },
    ...
  },
  "updated": "2021-10-13T19:00:54.723Z"
}
```

Also, adds a `dev:hotswap` command that runs `cdk deploy --hotswap` for faster lambda deployment. (Note that the lambda must still be locally bundled beforehand.)

Builds off of the changes in #372 

Closes #369, #370, #373, #374

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*