chore: tweak eligibility detection & catalog builder
Tweaked the eligibility detection rules to more accurately capture
`cdktf` packages, and tuned the framework detection accordinly (to have
a better coverage).

Additionally, updated the catalog builder to better handle cases where
a package is re-processed for a metadata update, but no new version is
added. Previously, the catalog would only be updated if the new package
version was strictly greater than the current one; and this changes it
to also update the contents if the new version is equal to the currently
registered one.