chore: tweak eligibility detection & catalog builder (#397)
Tweaked the eligibility detection rules to more accurately capture
`cdktf` packages, and tuned the framework detection accordinly (to have
a better coverage).

Additionally, updated the catalog builder to better handle cases where
a package is re-processed for a metadata update, but no new version is
added. Previously, the catalog would only be updated if the new package
version was strictly greater than the current one; and this changes it
to also update the contents if the new version is equal to the currently
registered one.

Finally, the catalog builder function, when operating in "rebuild" mode,
will continue in a separate execution when it has less than 1 minute
remaining before timeout while traversing the S3 bucket. This was
introduced as the previous implementation could run for more than 15
minutes, which results in a timeout, and no work being saved.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*