fix: catalog rebuild truncates the catalog
The catalog builder was not properly registering packages found in an
existing catalog when operating in "refresh" or "continuation" mode, due
to excessive reliance on code from the "from scratch" mode only.

Edited the test case to actually be "refresh" instead of "rebuild" to
exercize the new code path properly.