chore: revert "chore: clean up CloudFront returned headers (#398)" (#402)
This reverts commit 6e374efe479cea0ff2f7d4d30127c43dff65a38e. It caused the website to stop returning responses, with the error message:

```
503 ERROR
The request could not be satisfied.
The CloudFront function associated with the CloudFront distribution is invalid or could not run. We can't connect to the server for this app or website at this time. There might be too much traffic or a configuration error. Try again later, or contact the app or website owner.
If you provide content to customers through CloudFront, you can find steps to troubleshoot and help prevent this error by reviewing the CloudFront documentation. 
```

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*