fix: NpmJS follower times out / appears to hang
The NpmJS follower is requesting batches of 100 changes by default, but
this can result in excessively large response payloads. There is a limit
in how long a string can be when created from a `Buffer` instance, and
`JSON.parse` only operates on `string` instances.

Consequently, we aim to detect when the response has grown larger than
what we can stringify, and throw a dedicated error in case this happens.
The `changes` endpoint detects this particular error condition and
retries with a batch size reduced by half until the batch size is equal
to 1, at which point it will let the error bubble up to the client.

The follower appeared to hang because the `close` event of the HTTP
message was crashing without ever calling the `resolve` nor `reject`
handlers of the `Promise`, because it was throwing an error before
getting there.