feat: set explicit Cache-Control headers
On all S3 objects that are served by CloudFront, set explicit
Cache-Control headers (typically with a 5 minutes expiration,
must-revalidate and proxy-revalidate, but occasionally with a longer
client cache TTL for files that are not supposed to change often if at
all).

This removes the need to perform explicit cache invaldiations on the
CloudFront distribution, which can be problematic, as there can only
be a few concurrent invalidations, and that can cause issues during
deployments.