fix: SFN fails on ECS.InvalidParameterException
When ECS gets throttled when trying to obtain security group information,
it results in a `ECS.InvalidParameterException` error in StepFunctions,
with an error message that mentions the throttle event.

Adds the `ECS.InvalidParameterException` error to the retry logic, and
basically treat it as a synonym of `ECS.AmazonECSException` in this
context.