chore: rewrite generated ECS entry-point in TypeScript
Ensure the entry-point gets type-checked, linted, and down-leveled by
`esbuild` so that it's generally easier to ensure it behaves as
intended.

A side-effect of this change is that the generated handler is now
checked-in, which also makes it's content easier to code-review.

Finally, this switches the node version in use by the ECS container
images to Node 16.x, which provides somewhat better syntax support, and
offers a better stack trace experience in asynchronous contexts (which
is our case here).