fix: CodeArtifact may be behind npm, causing ETARGET in doc-gen
Add long-pause retries around `ETARGET` errors encountered when
attempting to install packages, as this can be a sign that the package
is not yet visible in CodeArtifact (may take 10 minutes or so), or that
one of it's dependencies was published "simultaneously" and isn't
visible yet (either because npm doesn't have it yet, or because
CodeArtifact does not show it yet).

This is limited to 3 retries, with 5, 10 then 20 minutes back-off (for
a grand total of 35 minutes of back-off). The mesage will hit the
dead-letter queue if it hasn't resolved by this time.