fix: move ingestion configuration to external file
Moves the `packageLinks` and `packageTags` configuration from being
passed to the ingestion function via environment variables, to being
written to a file in a s3 bucket. This allows the configuration to be
much larger where previously it was easy to cause the deploy to fail
because environment variables have a limit of 5120 bytes.

Abstracts out the `ConfigFile` class which takes a filename and a string
and writes the file to a temporary directory to allow usage in cdk
assets. Used by the web app configuration and ingestion fn
configuration.

Update ingestion unit tests to test consumption of configuration via s3.