chore: generate all documentation in a single ECS task
Instead of generating each language in a separate task, run a sigle task
to generate all languages.

This has the advantage of reducing the amount of ECS containers being
created, and the amount of packages installed from CodeArtifact, by a
factor of 4 (since we currently generate 4 different languages).

This means iether all languages succeed, or all fail, which is less
flexible than what we did before, however, we can observe that usually,
packages fail at `npm install`, and in such cases the failure is shared
across all languages... The flexibility does not buy us much.

We are also in control of the timeout of the ECS tasks (these can run
for extended periods of time), so we can cope with the future increase
of language count.