feat: use separate overlay repository for self-publishing
Instead of self-publishing to the lowest-order repository, publish to
a no-upstream, no-external-connections overlay-only repository that is
the last upstream to the repository used for installing.

This removes the risk of interference from the external connection, and
allows us to give higher precedence to "externally" obtained packages
when they are available.