feat: esbuild watch commands (#458)
This PR adds a suite of esbuild watch commands to make it easier to continuously build lambda code for development iteration speed (which can easily be used with the cdk hotswap command available via `dev:hotswap`. Individual lambdas can be watched by calling e.g. `bundle:npm-js-follower:watch`, and all lambdas can be watched together with `bundle:lambda:watch`.

I wasn't able to add a unified watch command for lambdas and fargate tasks in this PR because at the moment the esbuild tasks are targeting different versions of node.

The trick to getting the "watch all" commands to work with our file structure is figuring out a way so that `xxx.task.ts` files in the same directory still end up with their bundles in distinct directories. This is achieved through the `--entry-names="[dir]/[name].bundle/index"` flag.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*