chore: speed up orchestration workflow for re-processing
Skip the "add to catalog" step unless an update to the catalog object is
likely useful (which can be checked parallelly). This reduces the
contention on the catalog builder function (concurrency of 1), and
allows workflows to complete faster.

This helps get a complete signal that re-processing has been successful
faster since workflows will no longer need to queue up in front of the
catalog builder, for possibly large amounts of time.

This is particularly precious under burst situations.