feat: custom CloudWatch dashboard widget for missing documentation
Replace the graphs for misisng package versions/submodules documentation
objects with a custom widget that shows the list of package versions
that are missing at least one documentation object for the given
language.

This involves an update to the inventory canary function, such that it
now produces new objects under the `missing-objects/` key prefix, which
contain the list of package versions (as a JSON-encoded array of
strings), which is then leveraged by the custom widget.

The widget itself only displays up to 100 individual packages (possibly
several versions for each of these), in order to avoid generating a
table that could be large enough to cause browsers to struggle rendering
the dashboard. In such cases, a link to the S3 object is included.

Fixes cdklabs/construct-hub-webapp#534