feat: support limited internet access mode
Introduced a new isolation mode that allows ConstructHub to access
internet, but only on certain CIDR ranges (those of CloudFlare - which
fronts `registry.npmjs.com`, and some of the ranges from GitHub).

This allows removing the hard dependency on CodeArtifact while
maintaining a balanced security posture around downloading & installing
third-party artifacts for documentation-generation purposes.

The CIDR lists are automatically kept up-to-date by GitHub workflows,
triggering on a daily basis, which will cut a PR to update the CIDR
allow-list files whenever they change.

The network restrictions are implemented by way of a series of
SecurityGroups (we need multiple groups, as these are limited to 60
rules each, and there are more than 60 CIDR blocks today), which are
registered on the compute tasks that perform work that is deemed "risky".