fix: npmjs package canary failed on HEAD requests
The `.end()` call is necessary when using `https.request()`, but was
missing, which eventually resulted in the connection being aborted by
the service or firewall, which triggered an un-caught `ECONNRESET`
error.

This caused the `node` VM to abort, and ultimately prevented the updated
state object to be written to S3 in the `finally` clause.