fix: type links are broken (#569)
The type links construct-hub creates are broken because they still format links using the old `#` pattern. 

We dropped this in https://github.com/cdklabs/construct-hub-webapp/pull/611 but didn't adjust the link formatter we pass to docgen to account for it.

This video shows the experience with links now. Note that clicking on a type link now reloads the page, which is worse experience that we used to have, and what we currently have with readme links. 

https://user-images.githubusercontent.com/1428812/142417380-d4045665-c775-4ae6-a491-3bce2381743a.mov

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*