fix: inconsistent snapshot tests (#587)
Asset hashes are inconsistent when the contents of package.json is
changed. This breaks the release flow because the version is changed
before the build runs which causes the contents of the webapp config
asset to differ.

Adds a mock package.json file that includes fields used by webapp
config. Maps this mock to the relevant module via jest options in
`.projenrc.js`.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*