fix(csp): re-unify the Content-Security-Policy headers
Stop serving the `Clear-Site-Data` header (as if honored, it would
result in clearing local storage, which is user-hostile as it'll nuke
the user's preferred language, etc...).

Re-unify the CSP-setting CloudFront function so that a consistent CSP is
served, instead of not serving one (or serving an incorrect one) for
select pages.