fix: set NpmJs follower alarm time grain to 5 minutes
The NpmJs follower alarms had a grain matchin the metrics' best
granularity (1 minute), however while the follower can produce several
points during an execution, it only runs every 5 minutes.

In order for the alarms to be reliable, the alarming time grain cannot
be lower than the periodicity of the follower invocations.