fix: unable to deploy updated IP allow-lists (#652)
We have provisions to cause the `CfnPrefixList` resource to be replaced
when its allow-list changes (the CIDR range capacity of `CfnPrefixList`
cannot be updated once it's deployed).

Due to how SecurityGroup in-/e-gress rules are managed by CDK, this
means the SecurityGroup that uses a `CfnPrefixList` will first be
updated by CloudFormation to add the new `CfnPrefixList` reference, and
during `CLEAN_UP`, the outdated list will be removed. This has the
potential of exceeding SecurityGroup rule count limits relatively
easily.

In order to address this problem, this causes the SecurityGroup to be
replaced whenever the `CfnPrefixList` gets replaced. As the
SecurityGroups are only used by Fargate task definitions, this will
cause a new revision of the task definition to be created, and will
hence not risk hitting the maximum number of SecurityGroups bound to a
network interface.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*