feat: version tracker (#637)
Closes https://github.com/cdklabs/construct-hub-webapp/issues/538
Closes https://github.com/cdklabs/construct-hub-webapp/issues/539
Closes https://github.com/cdklabs/construct-hub-webapp/issues/540

This adds a backend component that generates `all-versions.json`, a file in the backend that tracks all of the versions available of all packages, so that available version information can be displayed on the front end (to be added in a follow-up PR).

The file looks a bit like this:

```
{"packages": {"@aws-amplify/cdk-exported-backend": ["0.0.1","0.0.2","0.0.3","0.0.4","0.0.5"], "@aws-c2a/cdk-pipelines-step": ["0.7.1"], "@aws-cdk/assertions-alpha": ["2.0.0-alpha.1","2.0.0-alpha.2","2.0.0-alpha.3","2.0.0-alpha.4","2.0.0-rc.23","2.0.0-rc.24"], ....
```

To generate it, every minute a lambda makes ~1,000 S3 ListObjects calls in parallel over the course of 3-5 seconds*, and saves the results to the S3 bucket where package data is stored. The resulting JSON file is around 500kb on my development Construct Hub instance, or around half of the size of the catalog.

There are several other ways we could go about storing and updating this version data (e.g. storing a separate file per package, only updating the list of versions every time a new package is ingested, etc.) but these tend to add additional complexity since either more files need to be managed, or more failure modes are introduced as the component would have to integrate with the deny-list and ingestion systems.

Lambda invocations and S3 API calls aren't free, but my back-of-the-envelope calculation looked a bit like this:

```
S3 API calls: $7.2 / day = $0.000005 (per S3 requests) * 1000 (S3 requests per Lambda invoke) * (60 invokes / hr) * 24 (hrs / day)
Lambda cost:  $1.2 / day = $0.0000001667 (price per ms) * 5000 (ms per invoke) * 60 (invokes / hr) * 24 (hrs / day)
Total cost:   $8.4 / day
```

*These numbers are based on the ~800 packages we currently have in the public Construct Hub, and would be lower for private construct hub instances.

TODO:

- [ ] Add unit tests
- [ ] Add dashboard graphs

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*